Update main.c
解决GCC编译器警告"%d"与无符号长整型size_t类型不匹配
```
main.c:42:20: warning: format ‘%d’ expects argument of type ‘int’, but argument 2 has type ‘size_t {aka const long unsigned int}’ [-Wformat=]
             printf("ERROR-1: data length=%d which is not added yet; block size=%d\n", in->in_data_len, bs);
                    ^
main.c:42:20: warning: format ‘%d’ expects argument of type ‘int’, but argument 3 has type ‘size_t {aka const long unsigned int}’ [-Wformat=]
```

+++ src/main.c
@@ -39,7 +39,7 @@ void test_encrypt_with_cipher(const test_case_t *in, const EVP_CIPHER *cipher)
 		const size_t bs = EVP_CIPHER_block_size(cipher);
 		if (in->in_data_len % bs != 0)
 		{
-			printf("ERROR-1: data length=%d which is not added yet; block size=%d\n", in->in_data_len, bs);
+			printf("ERROR-1: data length=%d which is not added yet; block size=%d\n", (int) in->in_data_len, (int) bs);
 			/* Warning: Remember to do some clean-ups */
 			EVP_CIPHER_CTX_free(ctx);
 			return;
